+++ cli/cli/install.sh
@@ -1,6 +1,6 @@
 #bin/sh
 
 dotnet pack /p:Version=0.0.0
-dotnet tool uninstall beamcli -g || true
+dotnet tool uninstall beamable.tools -g || true
 dotnet tool uninstall cli -g || true
-dotnet tool install --global --add-source ./nupkg/ BeamCli
+dotnet tool install --global --add-source ./nupkg/ beamable.tools